Fix #354, Split shell API for optional implement
Defaults to not include OS_ShellOutputToFile implementation
