+++ src/os/portable/os-impl-no-shell.c
@@ -1,0 +1,30 @@
+/*
+ *      Copyright (c) 2018, United States government as represented by the
+ *      administrator of the National Aeronautics Space Administration.
+ *      All rights reserved. This software was created at NASA Glenn
+ *      Research Center pursuant to government contracts.
+ *
+ *      This is governed by the NASA Open Source Agreement and may be used,
+ *      distributed and modified only according to the terms of that agreement.
+ */
+
+#include "osapi.h"
+
+/**
+ * \file   os-impl-no-shell.c
+ *
+ * Purpose: No shell implementation, returns OS_ERR_NOT_IMPLEMENTED for calls
+ */
+
+/*----------------------------------------------------------------
+ *
+ * Function: OS_ShellOutputToFile_Impl
+ *
+ *  Purpose: Implemented per internal OSAL API
+ *           See description in os-impl.h for argument/return detail
+ *
+ *-----------------------------------------------------------------*/
+int32 OS_ShellOutputToFile_Impl(uint32 file_id, const char* Cmd)
+{
+    return OS_ERR_NOT_IMPLEMENTED;
+}
