+++ src/os/vxworks/osfileapi.c
@@ -25,13 +25,9 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <dirent.h>
-#include <errno.h>
 #include <sys/stat.h>
 #include <sys/types.h>
-#include <taskLib.h>
 #include <sysLib.h>
-#include <shellLib.h>
-#include <errnoLib.h>
 
 /* 
  * Leverage the POSIX-style File I/O as this will also work on vxWorks.
@@ -219,18 +215,6 @@ int32 OS_DirRemove_Impl(const char *local_path)
 } /* end OS_DirRemove_Impl */
 
 
-/****************************************************************************************
-                                     DEFINES
-****************************************************************************************/
-
-#define OS_REDIRECTSTRSIZE           15
-#define OS_SHELL_TMP_FILE_EXT        ".out"
-#define OS_SHELL_TMP_FILE_EXT_LEN    4
-#define OS_SHELL_CMD_TASK_STACK_SIZE 16384
-#define OS_SHELL_CMD_TASK_PRIORITY   250
-
-
-                        
 /*----------------------------------------------------------------
  *
  * Function: OS_VxWorks_StreamAPI_Impl_Init
@@ -267,62 +251,4 @@ int32 OS_VxWorks_DirAPI_Impl_Init(void)
    memset(OS_impl_dir_table, 0, sizeof(OS_impl_dir_table));
    return OS_SUCCESS;
 } /* end OS_VxWorks_DirAPI_Impl_Init */
-
-
-                        
-/*----------------------------------------------------------------
- *
- * Function: OS_ShellOutputToFile_Impl
- *
- *  Purpose: Implemented per internal OSAL API
- *           See prototype in os-impl.h for argument/return detail
- *
- *-----------------------------------------------------------------*/
-int32 OS_ShellOutputToFile_Impl(uint32 file_id, const char *Cmd)
-{
-    int32 ReturnCode = OS_FS_ERROR;
-    int32 Result = ERROR;
-    int32 fdCmd;
-    uint32 cmdidx;
-    char * shellName;
-
-    /* Create a file to write the command to (or write over the old one) */
-    fdCmd = OS_creat(OS_SHELL_CMD_INPUT_FILE_NAME,OS_READ_WRITE);
-
-    if (fdCmd < OS_FS_SUCCESS)
-    {
-        return OS_FS_ERROR;
-    }
-
-    if (OS_ConvertToArrayIndex(fdCmd, &cmdidx) == OS_SUCCESS)
-    {
-        /* copy the command to the file, and then seek back to the beginning of the file */
-        OS_write(fdCmd, Cmd, strlen(Cmd));
-        OS_lseek(fdCmd,0,OS_SEEK_SET);
-
-        /* Create a shell task the will run the command in the file, push output to OS_fd */
-        Result = shellGenericInit("INTERPRETER=Cmd", 0, NULL, &shellName, false, false,
-                OS_impl_filehandle_table[cmdidx].fd,
-                OS_impl_filehandle_table[file_id].fd,
-                OS_impl_filehandle_table[file_id].fd);
-    }
-
-    if (Result == OK)
-    {
-        /* Wait for the command to terminate */
-        do
-        {
-            taskDelay(sysClkRateGet());
-        }
-        while (taskNameToId(shellName) != ((TASK_ID)ERROR));
-
-        ReturnCode =  OS_SUCCESS;
-    }
-
-    /* Close the file descriptor */
-    OS_close(fdCmd);
-
-    return ReturnCode;
-
-} /* end OS_ShellOutputToFile_Impl */
