+++ src/unit-test-coverage/vxworks/modules/src/ut-osshell.c
@@ -1,0 +1,21 @@
+/*
+ *      Copyright (c) 2019, United States government as represented by the
+ *      administrator of the National Aeronautics Space Administration.
+ *      All rights reserved. This software was created at NASA Goddard
+ *      Space Flight Center pursuant to government contracts.
+ *
+ *      This is governed by the NASA Open Source Agreement and may be used,
+ *      distributed and modified only according to the terms of that agreement.
+ */
+
+/* pull in the OSAL configuration */
+#include "osconfig.h"
+#include "ut-osfileapi.h"
+
+/*
+ * Now include all extra logic required to stub-out subsequent calls to
+ * library functions and replace with our own
+ */
+#include "osshell.c"
+
+OS_VxWorks_filehandle_entry_t OS_impl_filehandle_table[OS_MAX_NUM_OPEN_FILES];
