+++ src/unit-test-coverage/vxworks/src/coveragetest-osfileapi.c
@@ -25,8 +25,6 @@
 #include "os-vxworks-coveragetest.h"
 
 #include <OCS_stdlib.h>
-#include <OCS_taskLib.h>
-#include <OCS_shellLib.h>
 #include <OCS_dirent.h>
 #include <OCS_unistd.h>
 #include <OCS_stat.h>
@@ -120,34 +118,6 @@ void Test_OS_DirRemove_Impl(void)
     OSAPI_TEST_FUNCTION_RC(OS_DirRemove_Impl("dir"), OS_ERROR);
 }
 
-void Test_OS_ShellOutputToFile_Impl(void)
-{
-    /*
-     * Test Case For:
-     * int32 OS_ShellOutputToFile_Impl(uint32 file_id, const char *Cmd)
-     */
-    int32 expected = OS_SUCCESS;
-    int32 actual;
-
-    /*
-     * The ShellOutputToFile will loop until the
-     * taskNameToId() function returns ERROR, so this
-     * must be set to avoid getting into an endless loop.
-     */
-    UT_SetDeferredRetcode(UT_KEY(OCS_taskNameToId), 2, -1);
-
-    actual = OS_ShellOutputToFile_Impl(0, "TestCmd");
-
-    UtAssert_True(actual == expected, "OS_ShellOutputToFile_Impl() (%ld) == OS_SUCCESS", (long)actual);
-    UtAssert_True(UT_GetStubCount(UT_KEY(OCS_shellGenericInit)) == 1, "shellGenericInit() called");
-
-    /* failure to open the output file */
-    UT_SetForceFail(UT_KEY(OS_creat), OS_ERROR);
-    expected = OS_ERROR;
-    actual = OS_ShellOutputToFile_Impl(0, "TestCmd");
-    UtAssert_True(actual == expected, "OS_ShellOutputToFile_Impl() (%ld) == OS_ERROR", (long)actual);
-}
-
 
 /* ------------------- End of test cases --------------------------------------*/
 
@@ -187,7 +157,6 @@ void UtTest_Setup(void)
     ADD_TEST(OS_DirRead_Impl);
     ADD_TEST(OS_DirRewind_Impl);
     ADD_TEST(OS_DirRemove_Impl);
-    ADD_TEST(OS_ShellOutputToFile_Impl);
 }
